refactor: LineGenerator 삭제 및 Ladder 생성 로직 변경
- LineGenerator 인터페이스 및 구현체 삭제, RandomLadderGenerator 로 대체
- Ladder의 생성 로직을 RandomLadderGenerator에 분리
- Ladder 생성 시 Line 리스트를 받는 방식으로 수정
