--- src/main/java/model/line/LineGenerator.java
@@ -1,6 +1,0 @@
-package model.line;
-
-public interface LineGenerator {
-
-    Line generateLine(int width);
-}
