+++ src/test/java/model/ladder/LadderTest.java
@@ -2,8 +2,6 @@ package model.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import model.line.LineGenerator;
-import model.line.RandomLineGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,9 +12,7 @@ class LadderTest {
     void testSizeOfLadderLines() {
         LadderHeight ladderHeight = new LadderHeight(5);
         LadderWidth ladderWidth = LadderWidth.from(4);
-        LineGenerator lineGenerator = new RandomLineGenerator();
-
-        Ladder ladder = Ladder.of(ladderHeight, ladderWidth, lineGenerator);
+        Ladder ladder = RandomLadderGenerator.generateLadder(ladderHeight, ladderWidth);
 
         assertThat(ladder.getLines()).hasSize(5);
     }
